some extra casting around size_t

+++ src/TinyGsmClientA6.h
@@ -581,7 +581,7 @@ protected:
   }
 
   int16_t modemSend(const void* buff, size_t len, uint8_t mux) {
-    sendAT(GF("+CIPSEND="), mux, ',', len);
+    sendAT(GF("+CIPSEND="), mux, ',', (uint16_t)len);
     if (waitResponse(2000L, GF(GSM_NL ">")) != 1) {
       return 0;
     }
