+++ src/TinyGsmClientM590.h
@@ -459,7 +459,7 @@ protected:
   }
 
   int16_t modemSend(const void* buff, size_t len, uint8_t mux) {
-    sendAT(GF("+TCPSEND="), mux, ',', len);
+    sendAT(GF("+TCPSEND="), mux, ',', (uint16_t)len);
     if (waitResponse(GF(">")) != 1) {
       return 0;
     }
